# swagger/server/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Накопление и отображение метрик.",
        "title": "Get All Metrics",
        "contact": {
            "email": "gam6itko@yandex.ru"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Получить все накопленные метрики в формате HTML.",
                "operationId": "GetAllMetrics",
                "responses": {
                    "200": {
                        "description": "Метрики",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/update/{type}/{name}/{value}": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Store"
                ],
                "summary": "Сохранить одну метрику.",
                "operationId": "UpdateOne",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Metric typ [counter, gauge]",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Metric name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Value",
                        "name": "value",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Метрика сохранена",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}